chore: rename rule to no-interpolation-in-snapshots
