+++ src/rules/no-interpolation-in-snapshots.ts
